add homework

+++ src/services/contacts/deleteContact.js
@@ -4,18 +4,21 @@ import { deleteFileFromCloudinary } from '../../utils/cloudinaryServices.js';
 import { deleteFileFromUploadDir } from '../../utils/uploadDirServices.js';
 
 export const deleteContact = async ({ contactId, userId }) => {
-  const { photo } = await Contact.findOne({
+  const contact = await Contact.findOne({
     _id: contactId,
     userId,
   });
+
+  if (!contact) return;
+
   const result = await Contact.findOneAndDelete({
     _id: contactId,
     userId,
   });
 
-  if (photo) {
+  if (contact && contact.photo !== null) {
     if (getEnvVar('ENABLE_CLOUDINARY') === 'true') {
-      const splittedUrl = photo.split('/');
+      const splittedUrl = contact.photo.split('/');
       const photoPublicId = splittedUrl[splittedUrl.length - 1].replace(
         '.jpg',
         '',
@@ -23,7 +26,7 @@ export const deleteContact = async ({ contactId, userId }) => {
       await deleteFileFromCloudinary(photoPublicId);
     }
   } else {
-    await deleteFileFromUploadDir(photo);
+    await deleteFileFromUploadDir(contact.photo);
   }
 
   return result;
